+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/result/BaseSearchResult.kt
@@ -48,6 +48,9 @@ abstract class BaseSearchResult(
     open val categories: List<String>?
         get() = rawSearchResult.categories
 
+    open val categoryIds: List<String>?
+        get() = rawSearchResult.categoryIds
+
     open val makiIcon: String?
         get() = rawSearchResult.icon